Do not write file if contents have not changed
Currently, any call of `ConfigUpdater#update` will write the new file with the computed contents. However, writting the file to the disk is NOT necessary if the "new" and "current" contents are the same. This commit changes this, by writting "updated contents" to a string first and then to the file if it detects any change.